+++ backend/src/main/java/com/unpaz/backend/repository/TicketRepository.java
@@ -1,9 +1,11 @@
 package com.unpaz.backend.repository;
 
 import com.unpaz.backend.model.Ticket;
+
+
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface TicketRepository extends JpaRepository<Ticket, Integer> {
+public interface TicketRepository extends JpaRepository<Ticket, Long> {
 }
